requirements.txt
streamlit
nltk
langchain
openai
python-dotenv